Refactor test folder, removed unnecesary components
