--- App/Components/Styles/README.md
@@ -1,2 +1,0 @@
-### Styles Folder
-Component styles are separated from functionality.
