+++ App/Redux/index.js
@@ -5,9 +5,7 @@ import rootSaga from '../sagas/'
 export default () => {
   /* ------------- Assemble The Reducers ------------- */
   const rootReducer = combineReducers({
-    github: require('./github_redux').reducer,
-    login: require('./login_redux').reducer,
-    search: require('./search_redux').reducer
+    github: require('./github_redux').reducer
   })
 
   return configureStore(rootReducer, rootSaga)
