+++ App/Sagas/index.js
@@ -5,14 +5,10 @@ import DebugConfig from '../config/debug_config'
 
 /* ------------- Types ------------- */
 
-import { StartupTypes } from '../redux/startup_redux'
 import { GithubTypes } from '../redux/github_redux'
-import { LoginTypes } from '../redux/login_redux'
 
 /* ------------- Sagas ------------- */
 
-import { startup } from './startup_sagas'
-import { login } from './login_sagas'
 import { getUserAvatar } from './github_sagas'
 
 /* ------------- API ------------- */
@@ -25,10 +21,6 @@ const api = DebugConfig.useFixtures ? FixtureAPI : API.create()
 
 export default function * root () {
   yield [
-    // some sagas only receive an action
-    takeLatest(StartupTypes.STARTUP, startup),
-    takeLatest(LoginTypes.LOGIN_REQUEST, login),
-
     // some sagas receive extra parameters in addition to an action
     takeLatest(GithubTypes.USER_REQUEST, getUserAvatar, api)
   ]
