+++ app/components/launch_panel/index.js
@@ -1,0 +1,3 @@
+import LaunchPanel from './launch_panel'
+
+export default LaunchPanel
