+++ app/components/launch_panel/launch_panel_styles.js
@@ -1,5 +1,5 @@
 import { StyleSheet } from 'react-native'
-import { Metrics, ApplicationStyles } from '../../themes/'
+import { Metrics, ApplicationStyles } from './../../themes'
 
 export default StyleSheet.create({
   ...ApplicationStyles.screen,
