+++ app/components/rounded_button/index.js
@@ -1,0 +1,3 @@
+import RoundedButton from './rounded_button'
+
+export default RoundedButton
