--- app/components/search_bar.js
@@ -1,41 +1,0 @@
-import React from 'react'
-import { View, Text, TextInput, TouchableOpacity } from 'react-native'
-import styles from './styles/search_bar_styles'
-import { Colors, Metrics } from '../themes/'
-import Icon from 'react-native-vector-icons/FontAwesome'
-
-export default class SearchBar extends React.Component {
-  static propTypes = {
-    onSearch: React.PropTypes.func.isRequired,
-    onCancel: React.PropTypes.func.isRequired,
-    searchTerm: React.PropTypes.string
-  }
-
-  render () {
-    const { onSearch, onCancel, searchTerm } = this.props
-    const onSubmitEditing = () => onSearch(searchTerm)
-    return (
-      <View style={styles.container}>
-        <Icon name='search' size={Metrics.icons.tiny} style={styles.searchIcon} />
-        <TextInput
-          ref='searchText'
-          autoFocus
-          placeholder='Search'
-          placeholderTextColor={Colors.snow}
-          underlineColorAndroid='transparent'
-          style={styles.searchInput}
-          value={this.props.searchTerm}
-          onChangeText={onSearch}
-          autoCapitalize='none'
-          onSubmitEditing={onSubmitEditing}
-          returnKeyType={'search'}
-          autoCorrect={false}
-          selectionColor={Colors.snow}
-        />
-        <TouchableOpacity onPress={onCancel} style={styles.cancelButton}>
-          <Text style={styles.buttonLabel}>Cancel</Text>
-        </TouchableOpacity>
-      </View>
-    )
-  }
-}
