--- app/components/styles/drawer_button_styles.js
@@ -1,9 +1,0 @@
-import { Metrics, Colors, Fonts } from '../../themes'
-
-export default {
-  text: {
-    ...Fonts.style.h5,
-    color: Colors.snow,
-    marginVertical: Metrics.baseMargin
-  }
-}
