--- app/components/styles/full_button_styles.js
@@ -1,20 +1,0 @@
-import { StyleSheet } from 'react-native'
-import { Fonts, Colors } from '../../themes/'
-
-export default StyleSheet.create({
-  button: {
-    marginVertical: 5,
-    borderTopColor: Colors.fire,
-    borderBottomColor: Colors.bloodOrange,
-    borderTopWidth: 1,
-    borderBottomWidth: 1,
-    backgroundColor: Colors.ember
-  },
-  buttonText: {
-    margin: 18,
-    textAlign: 'center',
-    color: Colors.snow,
-    fontSize: Fonts.size.medium,
-    fontFamily: Fonts.type.bold
-  }
-})
