+++ app/navigation/navigation_styles.js
@@ -1,5 +1,5 @@
 import { StyleSheet } from 'react-native'
-import { Colors } from '../../themes/'
+import { Colors } from './../themes'
 
 export default StyleSheet.create({
   header: {
