--- app/redux/startup_redux.js
@@ -1,10 +1,0 @@
-import { createActions } from 'reduxsauce'
-
-/* ------------- Types and Action Creators ------------- */
-
-const { Types, Creators } = createActions({
-  startup: null
-})
-
-export const StartupTypes = Types
-export default Creators
