--- app/sagas/login_sagas.js
@@ -1,13 +1,0 @@
-import { put } from 'redux-saga/effects'
-import LoginActions from '../redux/login_redux'
-
-// attempts to login
-export function * login ({ username, password }) {
-  if (password === '') {
-    // dispatch failure
-    yield put(LoginActions.loginFailure('WRONG'))
-  } else {
-    // dispatch successful logins
-    yield put(LoginActions.loginSuccess(username))
-  }
-}
