+++ app/screens/README.md
@@ -1,0 +1,2 @@
+### Screens Folder
+All screens of the application are stored and organized here
