+++ app/screens/launch_screen.js
@@ -1,0 +1,13 @@
+// Libs ========================================================================
+import React from 'react'
+
+// Components ==================================================================
+import LaunchContainer from './../containers/launch_container'
+
+export default class LaunchScreen extends React.Component {
+  render () {
+    return (
+      <LaunchContainer />
+    )
+  }
+}
