+++ app/services/rehydration_services.js
@@ -1,13 +1,12 @@
 import ReduxPersist from '../config/redux_persist'
 import { AsyncStorage } from 'react-native'
 import { persistStore } from 'redux-persist'
-import StartupActions from '../redux/startup_redux'
 import DebugConfig from '../config/debug_config'
 
 const updateReducers = (store: Object) => {
   const reducerVersion = ReduxPersist.reducerVersion
   const config = ReduxPersist.storeConfig
-  const startup = () => store.dispatch(StartupActions.startup())
+  // const startup = () => store.dispatch(StartupActions.startup())
 
   // Check to ensure latest reducer version
   AsyncStorage.getItem('reducerVersion')
@@ -25,14 +24,17 @@ const updateReducers = (store: Object) => {
           })
         }
         // Purge store
-        persistStore(store, config, startup).purge()
+        // persistStore(store, config, startup).purge()
+        persistStore(store, config).purge()
         AsyncStorage.setItem('reducerVersion', reducerVersion)
       } else {
-        persistStore(store, config, startup)
+        // persistStore(store, config, startup)
+        persistStore(store, config)
       }
     })
     .catch(() => {
-      persistStore(store, config, startup)
+      // persistStore(store, config, startup)
+      persistStore(store, config)
       AsyncStorage.setItem('reducerVersion', reducerVersion)
     })
 }
