+++ tests/services/fixture_api_test.js
@@ -1,5 +1,5 @@
-import API from '../../App/Services/Api'
-import FixtureAPI from '../../App/Services/FixtureApi'
+import API from '../../app/services/api'
+import FixtureAPI from '../../app/services/fixture_api'
 import R from 'ramda'
 
 test('All fixtures map to actual API', () => {
@@ -13,7 +13,7 @@ test('All fixtures map to actual API', () => {
 })
 
 test('FixtureAPI getRate returns the right file', () => {
-  const expectedFile = require('../../App/Fixtures/rateLimit.json')
+  const expectedFile = require('../../app/fixtures/rate_limit.json')
 
   expect(FixtureAPI.getRate()).toEqual({
     ok: true,
@@ -22,7 +22,7 @@ test('FixtureAPI getRate returns the right file', () => {
 })
 
 test('FixtureAPI getUser returns the right file for gantman', () => {
-  const expectedFile = require('../../App/Fixtures/gantman.json')
+  const expectedFile = require('../../app/fixtures/gantman.json')
 
   expect(FixtureAPI.getUser('GantMan')).toEqual({
     ok: true,
@@ -31,7 +31,7 @@ test('FixtureAPI getUser returns the right file for gantman', () => {
 })
 
 test('FixtureAPI getUser returns the right file for skellock as default', () => {
-  const expectedFile = require('../../App/Fixtures/skellock.json')
+  const expectedFile = require('../../app/fixtures/skellock.json')
 
   expect(FixtureAPI.getUser('Whatever')).toEqual({
     ok: true,
